Split up Featherstone into kernels.py and main file
